pc_app/requirements.txt
PySide6>=6.7
hidapi>=0.14.0
pyserial>=3.5
pyinstaller>=6.0
vgamepad>=0.1.0
